+++ src/pages/index.ts
@@ -6,4 +6,4 @@ export { default as Quiz } from './Quiz/Quiz';
 export { default as QuizEdit } from './QuizEdit/QuizEdit';
 export { default as QuizFilter } from './QuizFilter/QuizFilter';
 export { default as FilterParse } from './QuizFilter/FilterParse';
-export { default as Auth } from './Auth/Auth';
+export { default as Login } from './Auth/Login';
